refactor: move bot AI data models to discordbot.typings.games
